codewriter: widen BhSizeSpec/BhDescr vtable to u64
The vtable field is an ob_type pointer captured in the producing process
(the opcode_descrs.bin build script) and crosses the build->runtime
serialization boundary. Declared usize, a 64-bit host pointer failed
bincode's width check when deserialized on a wasm32 runtime. Store it as
u64 and cast at the consumers; the value is an opaque identity word
either way (real vtables are re-resolved via type_id -> gc_cache).

+++ majit/majit-metainterp/src/pyjitpl/dispatch.rs
@@ -616,7 +616,7 @@ fn size_descr_ref_from_bh(descr: &crate::blackhole::BhDescr) -> majit_ir::DescrR
                 *size,
                 *type_id as u32,
                 *type_id,
-                *vtable,
+                *vtable as usize,
                 *is_gc_managed,
                 &specs,
             );
@@ -682,7 +682,7 @@ fn field_descr_ref_from_bh(descr: &crate::blackhole::BhDescr) -> (usize, majit_i
                         p.size,
                         p.type_id as u32,
                         p.type_id,
-                        p.vtable,
+                        p.vtable as usize,
                         p.is_gc_managed,
                         &specs,
                     );
